Sinatra switch to modular style
