+++ Gemfile
@@ -3,7 +3,7 @@ source "https://rubygems.org"
 ruby '2.2.3'
 
 gem 'rake'
-gem 'sinatra'
+gem 'sinatra', require: 'sinatra/base'
 gem 'sinatra-flash'
 gem 'sinatra-partial'
 gem 'bcrypt'
