+++ spec/helpers/session.rb
@@ -23,8 +23,3 @@ module SessionHelpers
 
 end
 
-
-
-
-
-
